+++ src/server/router/nodes.ts
@@ -1,13 +1,23 @@
 import { createRouter } from "./context";
 import { z } from "zod";
+import { TRPCError } from "@trpc/server";
 
 export const nodeRouter = createRouter()
+  .middleware(async ({ ctx, next }) => {
+    // Any queries or mutations after this middleware will
+    // raise an error unless there is a current session
+    if (!ctx.session) {
+      throw new TRPCError({ code: "UNAUTHORIZED" });
+    }
+    return next();
+  })
   .query("getAll", {
     input: z.object({userID: z.string().nullish()}),
     async resolve({ ctx, input }) {
-      if (input.userID === null)
-        return
-
+      if (input.userID === null) return;
+      if (input.userID !== ctx.session?.user?.id) {
+        throw new TRPCError({ code: "UNAUTHORIZED" });
+      }
       return await ctx.prisma.node.findMany({
         where: {
           userId: input.userID,
@@ -31,6 +41,9 @@ export const nodeRouter = createRouter()
     input: z.object({userID: z.string().nullish()}),
     async resolve({ ctx, input}) {
       if (input.userID === null) return;
+      if (input.userID !== ctx.session?.user?.id) {
+        throw new TRPCError({ code: "UNAUTHORIZED" });
+      }
       return await ctx.prisma.edge.findMany({
         select: {
           node1_id: true,
@@ -55,6 +68,9 @@ export const nodeRouter = createRouter()
       priority: z.string()
     }),
     async resolve({ input, ctx }) {
+      if (input.userId !== ctx.session?.user?.id) {
+        throw new TRPCError({ code: "UNAUTHORIZED" });
+      }
       return await ctx.prisma.node.upsert({
         where: {id: input.nodeId},
         update: {
@@ -84,8 +100,12 @@ export const nodeRouter = createRouter()
     input: z.object({
       nodeId: z.string(),
       layerId: z.string(),
+      userId: z.string(),
     }),
     async resolve({ input, ctx }) {
+      if (input.userId !== ctx.session?.user?.id) {
+        throw new TRPCError({ code: "UNAUTHORIZED" });
+      }
       return await ctx.prisma.nodeLayers.create({
         data: {
           nodeId: input.nodeId,
@@ -102,6 +122,9 @@ export const nodeRouter = createRouter()
     }),
     async resolve({ input, ctx }) {
       if (input.userId === null) return;
+      if (input.userId !== ctx.session?.user?.id) {
+        throw new TRPCError({ code: "UNAUTHORIZED" });
+      }
       return await ctx.prisma.edge.create({
         data: {
           node1_id: input.node1Id,
@@ -113,9 +136,13 @@ export const nodeRouter = createRouter()
   })
   .mutation("deleteNode", {
     input: z.object({
-      nodeId: z.string()
+      nodeId: z.string(),
+      userId: z.string()
     }),
     async resolve({ input, ctx }) {
+      if (input.userId !== ctx.session?.user?.id) {
+        throw new TRPCError({ code: "UNAUTHORIZED" });
+      }
       return await ctx.prisma.node.delete({
         where: {id: input.nodeId},
       });
@@ -124,9 +151,13 @@ export const nodeRouter = createRouter()
   .mutation("deleteNodeLayer", {
     input: z.object({
       nodeId: z.string(),
-      layerId: z.string()
+      layerId: z.string(),
+      userId: z.string()
     }),
     async resolve({ input, ctx }) {
+      if (input.userId !== ctx.session?.user?.id) {
+        throw new TRPCError({ code: "UNAUTHORIZED" });
+      }
       return await ctx.prisma.nodeLayers.deleteMany({
         where: {
           nodeId: input.nodeId,
@@ -138,9 +169,13 @@ export const nodeRouter = createRouter()
   .mutation("deletePair", {
     input: z.object({
       node1Id: z.string(),
-      node2Id: z.string()
+      node2Id: z.string(),
+      userId: z.string()
     }),
     async resolve({ input, ctx }) {
+      if (input.userId !== ctx.session?.user?.id) {
+        throw new TRPCError({ code: "UNAUTHORIZED" });
+      }
       return await ctx.prisma.edge.deleteMany({
         where: {
           node1_id: input.node1Id,
